Change Calculator.js & App.js to function component & remove initial from calculate.js
