+++ src/App.js
@@ -1,14 +1,19 @@
 import React from 'react';
 import Calculator from './components/Calculator';
 
+const App = () => <Calculator />;
+
+export default App;
+
+/*
 class App extends React.Component {
   constructor(props) {
     super(props);
     this.state = {};
   }
-
   render() {
     return <Calculator />;
   }
 }
 export default App;
+*/
